+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#883] Query tools `add`, `update`, `remove`, `move`, and `group` can now accept a `string` or `Path` where they previously accepted only `Path`-type arguments. A `string` would represent the `id` property of the subject rule/group, but otherwise the functionality of each tool is the same.
 - [#883] New utility methods: `findID` returns the rule or group with the given `id` property within a query hierarchy (similar to `findPath`), and `getPathOfID` returns the `path` of the rule or group with the given `id`.
 
+### Fixed
+
+- [#885] `parseSQL` was translating `NOT LIKE '[...]'` (no leading or trailing wildcards) to `operator: "="` instead of `operator: "!="`.
+
 ## [v8.5.0] - 2025-04-07
 
 ### Added